# Conflicts:
#	src/main/resources/public/js/guests.js

+++ src/main/resources/public/js/guests.js
@@ -67,13 +67,12 @@ function refreshTable() {
 }
 
 function del(id) {
-    $.alert({
-                title: 'Alert!',
-                content: 'Simple alert!',
-            });
-    $.ajax({url: "/api/removeGuest/" + id, type: "DELETE"}).done( function() {
-        refreshTable();
-    })
+    var r = confirm("Weet je zeker dat je de informatie van \"naam\" wilt verwijderen?");
+    if (r == true) {
+        $.ajax({url: "/api/removeGuest/" + id, type: "DELETE"}).done( function() {
+                refreshTable();
+            })
+        }
 }
 
 // When the user clicks on <div>, open the popup
